Structure API tests, tests for wallet recovery phrase and allow less verbosity with headless
